fix: details link in options vesting email (#1219)
When an invoice with an equity swap is paid, the worker user receives an
email detailing the vested option information along with a link to the
Equity Dashboard. However, this link incorrectly redirects the user to
`/equity/grants`, which is not accessible to worker / investors.

PR fixes the link to `/equity/options` which shows details related to
those options.

Part of #911