Support past investors as a user type (#1272)
### What

An investor who has sold all their securities in a company should be
able to view their past information, such as dividends and documents,
but not information that is kept up to date, like the cap table and the
Updates page.

Refer: #1115

### Why

The records should mainly be preserved in the system because
`company_investors.id` is referenced by multiple tables. Since we need
to do that, we need to have separation in the application logic to treat
active and inactive investors differently.

### Notes

AI usage: Cursor tab & Cursor AI Code Generation

- [x] Hide buybacks
- [x] Hide updates
- [x] Hide Cap table
- [x] Hide Equity main link if there are no sublinks
- [x] Don't send company updates to deactivated investors
- [x] Ensure they're ignored during dividend calculations
- [x] Don't send active-investor targeted emails (look for them)
- [x] Write tests
  - [x] Unit
  - [x] E2E
- [x] Manual test

---------

Co-authored-by: Sahil Lavingia <sahil.lavingia@gmail.com>