Send email on every invocation of `remind_dividend_investors` (#1409)
### What

Change the `DividendRound#remind_dividend_investors` method to send a
reminder email for a dividend, even if we had already emailed about it
during issuance.

### Why

This is the intended behavior, but it was incorrectly implemented. It
came to light in [another
PR](https://github.com/antiwork/flexile/pull/1309#discussion_r2394819390).
In the existing behavior, an invocation of the method does not send any
emails because `InvestorDividendRound#send_dividend_issued_email` only
sends an email once, regardless of how many times it is called.

## Notes

AI usage: none