Status column not sorting properly (#1398)
Closes #1396 

## AI Disclosure
No AI was used

## Root cause
The status column has a filtering based on 3 status' (`["Active",
"Onboarding", "Alumni"]`) and the same filtering status was used to sort
this column instead of using the values of the date fields. Therefore
the rows were being sorted randomly based on what status was assigned
(multiple rows could have same status).

## The fix
To fix this a `sortingFn` is passed to specify the behaviour of sorting
when clicking on the sort buttons

## Demo
Included demo to show both sorting and filtering working correctly


https://github.com/user-attachments/assets/1a2176a8-99f0-4e35-8439-79e57d9be1df


https://github.com/user-attachments/assets/d21d1bd3-b05c-4f61-8da5-55d2925a4891

### Mobile
No impact on the mobile view as sorting is not available on mobile

<img width="533" height="848" alt="image"
src="https://github.com/user-attachments/assets/b49a0eec-f974-4ecb-8d66-4ef81261cd06"
/>





## E2E
Added a new e2e test to verify this

---------

Co-authored-by: Kavitha809 <kanhaya.yadav@guardinex.com>
Co-authored-by: Sahil Lavingia <sahil.lavingia@gmail.com>